Split CustomLesson out of Lesson
